+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### 🐛 Fixes
 - Update use cases when a space between the number sign (#) and the number is removed. Before, the space was falsely removed between level 2 markdown headings and numbers (e.g. “## 1. introduction” was falsely changed to “##1. introduction”). Now the space is kept. (bug reported by [@filipaldi](https://github.com/filipaldi))
 - A dash was not fixed between words, when the first word ended on a letter “ŷ”
+- An nbsp was falsely added after a word that had the letter “ŷ” on the second to the last past position, e.g. nbsp was falsely added after the word “starŷm”
 
 
 ### 🔨 Maintenance